refactor:abstracted out changeColor function
