+++ src/components/notecard-component/NoteCard.jsx
@@ -5,6 +5,7 @@ import "./NoteCard.css";
 import { addOrRemoveFromArchive } from "../../utils/notes-util/addOrRemoveFromArchive";
 import { addToTrash } from "../../utils/trash-util/addToTrash";
 import { HexColorPicker } from "react-colorful";
+import { changeColor } from "../../utils/notes-util/changeColor";
 
 const deletePermanently=(trashNote,dispatchNotes)=>{
     dispatchNotes({type:"DELETE_NOTE",trashNote:trashNote})
@@ -15,25 +16,7 @@ const editHandler=(_id,dispatchNotes,createNote,setCreateNote)=>{
     setCreateNote(!createNote)
 }
 
-const changeColor=async(changedCol,note,dispatchNotes)=>{
-  const encodedToken=localStorage.getItem("token")
-  const {_id, title, description, tags, createdAt }=note
-  try {
-    const response=await axios({
-      method:"POST",
-      url:`/api/notes/${note._id}`,
-      headers:{authorization:encodedToken},
-      data:{note:{
-        _id, title, description, color:changedCol, tags, createdAt 
-      }}
-    })
-    if(response.status===201){
-      dispatchNotes({type:"SET_NOTES",payload:response.data.notes})
-    }
-  } catch (error) {
-    console.log('error occured',error)
-  }
-}
+
 
 function NoteCard({
   noteItem: { _id, title, description, color, tags, createdAt },
